em_django/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.6
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
